services/finance-advisors/requirements.txt
langgraph>=0.2.0
langchain>=0.2.16
langchain-community>=0.2.10
langchain-ollama>=0.1.0
ollama>=0.3.0
python-dotenv>=1.0.1
